Extend common Maze class
