+++ src/js/components/lib/cells/Door.js
@@ -10,9 +10,10 @@ class Door {
 
   draw(ctx, x, y, size) {
     ctx.beginPath();
+
     ctx.fillStyle = "rgba(112, 153, 77, 1)";
-    ctx.rect(x, y, size, size);
-    ctx.fill();
+    ctx.fillRect(x, y, size, size);
+
     ctx.closePath();
   }
 }
