+++ src/js/components/lib/cells/Wall.js
@@ -10,8 +10,10 @@ class Wall {
 
   draw(ctx, x, y, size) {
     ctx.beginPath();
+
     ctx.fillStyle = "rgba(111, 109, 109, 1)";
     ctx.fillRect(x, y, size, size);
+
     ctx.closePath();
   }
 }
